pyproject.toml
[tool.poetry]
name = "CargoService"
version = "0.1.0"
description = ""
authors = ["JMURv <architect.lock@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
geopy = "^2.3.0"
celery = "^5.2.7"
django-celery-beat = "^2.5.0"
redis = "^4.5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
